tests/wpt/tests/css/css-scoping/slot-non-html-display-value.html

<!doctype html>
<title>CSS Test: non-HTML slot elements don't get display: contents by default.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="help" href="https://drafts.csswg.org/css-scoping/#slots-in-shadow-tree">
<link rel="help" href="https://bugzil.la/1468127">
<body>
<script>
test(function() {
  const non_html_slot = document.createElementNS("http://www.w3.org/2000/svg", "slot");
  document.body.appendChild(non_html_slot);
  const display = getComputedStyle(non_html_slot).display;
  // contents would get turned to "none" due to https://drafts.csswg.org/css-display/#unbox-svg
  assert_true(display != "contents" && display != "none");
}, "Non-HTML slot elements shouldn't be display: contents by default.");
</script>
</body>
